fix(numpy): aliases of builtin types is deprecated as of numpy 1.20 (#634)

+++ autotest/test_gwt_mst01.py
@@ -88,7 +88,7 @@ def build_models():
         dis = flopy.mf6.ModflowGwfdis(gwf, nlay=nlay, nrow=nrow, ncol=ncol,
                                       delr=delr, delc=delc,
                                       top=top, botm=botm,
-                                      idomain=np.ones((nlay, nrow, ncol), dtype=np.int),
+                                      idomain=np.ones((nlay, nrow, ncol), dtype=int),
                                       filename='{}.dis'.format(gwfname))
 
         # initial conditions
@@ -209,7 +209,7 @@ def eval_transport(sim):
         assert False, 'could not load data from "{}"'.format(fpth)
 
     # end of stress period 1
-    cres1 = np.ones((nlay, nrow, ncol), np.float)
+    cres1 = np.ones((nlay, nrow, ncol), float)
     assert np.allclose(cres1, conc1), ('simulated concentrations do not match '
                                        'with known solution.')
